chore(outbound): Simplify router test utilities
In preparation for adding more tests, this commit simplifies the test setup
utilities so that they can be shared and reused across test modules.
