+++ linkerd/app/outbound/src/http/logical/tests/retries.rs
@@ -1,23 +1,18 @@
-use super::{
-    super::{policy, Outbound, ParentRef, Routes},
-    default_backend, http_get, mk_grpc_rsp, mk_rsp, send_req, serve_delayed, serve_req,
-    HttpConnect, Request, Response, Target,
-};
-use crate::test_util::*;
+use super::*;
 use hyper::body::HttpBody;
 use linkerd_app_core::{
     errors,
     proxy::http::{self, StatusCode},
-    svc::{self, http::stream_timeouts::StreamDeadlineError, NewService},
-    trace, NameAddr,
+    svc::http::stream_timeouts::StreamDeadlineError,
+    trace,
 };
 use linkerd_proxy_client_policy::{
     self as client_policy,
     grpc::{Codes, RouteParams as GrpcParams},
     http::{RouteParams as HttpParams, Timeouts},
 };
-use std::{collections::BTreeSet, net::SocketAddr, sync::Arc, time::Duration};
-use tokio::{sync::watch, time};
+use std::collections::BTreeSet;
+use tokio::time;
 use tonic::Code;
 use tracing::{info, Instrument};
 
@@ -25,7 +20,7 @@ use tracing::{info, Instrument};
 async fn http_5xx() {
     let _trace = trace::test::trace_init();
 
-    const TIMEOUT: Duration = Duration::from_secs(2);
+    const TIMEOUT: time::Duration = time::Duration::from_secs(2);
     let (svc, mut handle) = mock_http(HttpParams {
         timeouts: Timeouts {
             request: Some(TIMEOUT),
@@ -40,31 +35,29 @@ async fn http_5xx() {
         }),
     });
 
-    info!("Sending a request that will initially fail and then succeed");
     tokio::spawn(
         async move {
             handle.allow(2);
-            info!("Failing the first request");
-            serve_req(&mut handle, mk_rsp(StatusCode::INTERNAL_SERVER_ERROR, "")).await;
-            info!("Serving the second request");
-            serve_req(&mut handle, mk_rsp(StatusCode::NO_CONTENT, "")).await;
+            serve(&mut handle, mk_rsp(StatusCode::INTERNAL_SERVER_ERROR, "")).await;
+            serve(&mut handle, mk_rsp(StatusCode::NO_CONTENT, "")).await;
             handle
         }
         .in_current_span(),
     );
 
-    info!("Verifying that we see the successful response");
+    info!("Sending a request that will initially fail and then succeed");
     let rsp = time::timeout(TIMEOUT, send_req(svc.clone(), http_get()))
         .await
         .expect("response");
+    info!("Verifying that we see the successful response");
     assert_eq!(rsp.expect("response").status(), StatusCode::NO_CONTENT);
 }
 
 #[tokio::test(flavor = "current_thread", start_paused = true)]
-async fn http_5xx_limits() {
+async fn http_5xx_limited() {
     let _trace = trace::test::trace_init();
 
-    const TIMEOUT: Duration = Duration::from_secs(2);
+    const TIMEOUT: time::Duration = time::Duration::from_secs(2);
     let (svc, mut handle) = mock_http(HttpParams {
         timeouts: Timeouts {
             request: Some(TIMEOUT),
@@ -83,14 +76,26 @@ async fn http_5xx_limits() {
     tokio::spawn(
         async move {
             handle.allow(3);
-            info!("Failing the first request");
-            serve_req(&mut handle, mk_rsp(StatusCode::INTERNAL_SERVER_ERROR, "")).await;
-            info!("Failing the second request");
-            serve_req(&mut handle, mk_rsp(StatusCode::INTERNAL_SERVER_ERROR, "")).await;
-            info!("Failing the third request");
-            serve_req(&mut handle, mk_rsp(StatusCode::GATEWAY_TIMEOUT, "")).await;
+            serve(&mut handle, async move {
+                info!("Failing the first request");
+                mk_rsp(StatusCode::INTERNAL_SERVER_ERROR, "").await
+            })
+            .await;
+            serve(&mut handle, async move {
+                info!("Failing the second request");
+                mk_rsp(StatusCode::INTERNAL_SERVER_ERROR, "").await
+            })
+            .await;
+            serve(&mut handle, async move {
+                info!("Failing the third request");
+                mk_rsp(StatusCode::GATEWAY_TIMEOUT, "").await
+            })
+            .await;
             info!("Prepping the fourth request (shouldn't be served)");
-            serve_req(&mut handle, mk_rsp(StatusCode::NO_CONTENT, "")).await;
+            serve(&mut handle, async move {
+                mk_rsp(StatusCode::NO_CONTENT, "").await
+            })
+            .await;
             handle
         }
         .in_current_span(),
@@ -107,7 +112,7 @@ async fn http_5xx_limits() {
 async fn http_timeout() {
     let _trace = trace::test::trace_init();
 
-    const TIMEOUT: Duration = Duration::from_secs(2);
+    const TIMEOUT: time::Duration = time::Duration::from_secs(2);
     let (svc, mut handle) = mock_http(HttpParams {
         timeouts: Timeouts {
             request: Some(TIMEOUT),
@@ -127,16 +132,18 @@ async fn http_timeout() {
         async move {
             handle.allow(2);
 
-            info!("Delaying the first request");
-            serve_delayed(
-                TIMEOUT / 2,
-                &mut handle,
-                Ok(mk_rsp(StatusCode::NOT_FOUND, "")),
-            )
+            serve(&mut handle, async move {
+                info!("Delaying the first request");
+                time::sleep(TIMEOUT / 2).await;
+                mk_rsp(StatusCode::NOT_FOUND, "").await
+            })
             .await;
 
-            info!("Serving the second request");
-            serve_req(&mut handle, mk_rsp(StatusCode::NO_CONTENT, "")).await;
+            serve(&mut handle, async move {
+                info!("Serving the second request");
+                mk_rsp(StatusCode::NO_CONTENT, "").await
+            })
+            .await;
 
             handle
         }
@@ -154,7 +161,7 @@ async fn http_timeout() {
 async fn http_timeout_on_limit() {
     let _trace = trace::test::trace_init();
 
-    const TIMEOUT: Duration = Duration::from_secs(2);
+    const TIMEOUT: time::Duration = time::Duration::from_secs(2);
     let (svc, mut handle) = mock_http(HttpParams {
         timeouts: Timeouts {
             request: Some(TIMEOUT),
@@ -169,25 +176,22 @@ async fn http_timeout_on_limit() {
         }),
     });
 
-    info!("Testing that a retry timeout does not apply when max retries is reached");
     tokio::spawn(
         async move {
             handle.allow(2);
 
-            info!("Delaying the first request");
-            serve_delayed(
-                TIMEOUT / 3,
-                &mut handle,
-                Ok(mk_rsp(StatusCode::NOT_FOUND, "")),
-            )
+            serve(&mut handle, async move {
+                info!("Delaying the first request");
+                time::sleep(TIMEOUT / 3).await;
+                mk_rsp(StatusCode::NOT_FOUND, "").await
+            })
             .await;
 
-            info!("Delaying the second request");
-            serve_delayed(
-                TIMEOUT / 3,
-                &mut handle,
-                Ok(mk_rsp(StatusCode::NO_CONTENT, "")),
-            )
+            serve(&mut handle, async move {
+                info!("Delaying the second request");
+                time::sleep(TIMEOUT / 3).await;
+                mk_rsp(StatusCode::NO_CONTENT, "").await
+            })
             .await;
 
             handle
@@ -195,10 +199,12 @@ async fn http_timeout_on_limit() {
         .in_current_span(),
     );
 
-    info!("Verifying that the initial request was retried");
+    info!("Testing that a retry timeout does not apply when max retries is reached");
     let rsp = time::timeout(TIMEOUT, send_req(svc.clone(), http_get()))
         .await
         .expect("response");
+
+    info!("Verifying that the initial request was retried");
     assert_eq!(rsp.expect("response").status(), StatusCode::NO_CONTENT);
 }
 
@@ -206,7 +212,7 @@ async fn http_timeout_on_limit() {
 async fn http_timeout_with_request_timeout() {
     let _trace = trace::test::trace_init();
 
-    const TIMEOUT: Duration = Duration::from_millis(100);
+    const TIMEOUT: time::Duration = time::Duration::from_millis(100);
     let (svc, mut handle) = mock_http(HttpParams {
         timeouts: Timeouts {
             request: Some(TIMEOUT * 5),
@@ -226,47 +232,49 @@ async fn http_timeout_with_request_timeout() {
         async move {
             handle.allow(6);
 
-            info!("Delaying the first request");
-            serve_delayed(
-                TIMEOUT * 2,
-                &mut handle,
-                Ok(mk_rsp(StatusCode::IM_A_TEAPOT, "")),
-            )
+            // First request.
+
+            serve(&mut handle, async move {
+                info!("Delaying the first request");
+                time::sleep(TIMEOUT * 2).await;
+                mk_rsp(StatusCode::IM_A_TEAPOT, "").await
+            })
             .await;
 
-            info!("Delaying the second request");
-            serve_delayed(
-                TIMEOUT * 2,
-                &mut handle,
-                Ok(mk_rsp(StatusCode::IM_A_TEAPOT, "")),
-            )
+            serve(&mut handle, async move {
+                info!("Delaying the second request");
+                time::sleep(TIMEOUT * 2).await;
+                mk_rsp(StatusCode::IM_A_TEAPOT, "").await
+            })
             .await;
 
-            info!("Delaying the third request");
-            serve_req(&mut handle, mk_rsp(StatusCode::NO_CONTENT, "")).await;
-
-            info!("Delaying the fourth request");
-            serve_delayed(
-                TIMEOUT * 2,
-                &mut handle,
-                Ok(mk_rsp(StatusCode::IM_A_TEAPOT, "")),
-            )
+            serve(&mut handle, async move {
+                info!("Delaying the third request");
+                mk_rsp(StatusCode::NO_CONTENT, "").await
+            })
             .await;
 
-            info!("Delaying the fifth request");
-            serve_delayed(
-                TIMEOUT * 2,
-                &mut handle,
-                Ok(mk_rsp(StatusCode::IM_A_TEAPOT, "")),
-            )
+            // Second request
+
+            serve(&mut handle, async move {
+                info!("Delaying the fourth request");
+                time::sleep(TIMEOUT * 2).await;
+                mk_rsp(StatusCode::IM_A_TEAPOT, "").await
+            })
             .await;
 
-            info!("Delaying the sixth request");
-            serve_delayed(
-                TIMEOUT * 5,
-                &mut handle,
-                Ok(mk_rsp(StatusCode::NO_CONTENT, "")),
-            )
+            serve(&mut handle, async move {
+                info!("Delaying the fifth request");
+                time::sleep(TIMEOUT * 2).await;
+                mk_rsp(StatusCode::IM_A_TEAPOT, "").await
+            })
+            .await;
+
+            serve(&mut handle, async move {
+                info!("Delaying the sixth request");
+                time::sleep(TIMEOUT * 5).await;
+                mk_rsp(StatusCode::NO_CONTENT, "").await
+            })
             .await;
 
             handle
@@ -293,7 +301,7 @@ async fn http_timeout_with_request_timeout() {
 async fn grpc_internal() {
     let _trace = trace::test::with_default_filter("linkerd=debug");
 
-    const TIMEOUT: Duration = Duration::from_millis(100);
+    const TIMEOUT: time::Duration = time::Duration::from_millis(100);
     let (svc, mut handle) = mock_grpc(GrpcParams {
         timeouts: Timeouts {
             request: Some(TIMEOUT),
@@ -318,9 +326,9 @@ async fn grpc_internal() {
         async move {
             handle.allow(2);
             info!("Failing the first request");
-            serve_req(&mut handle, mk_grpc_rsp(tonic::Code::Internal)).await;
+            serve(&mut handle, mk_grpc_rsp(tonic::Code::Internal)).await;
             info!("Serving the second request");
-            serve_req(&mut handle, mk_grpc_rsp(tonic::Code::Ok)).await;
+            serve(&mut handle, mk_grpc_rsp(tonic::Code::Ok)).await;
             handle
         }
         .in_current_span(),
@@ -358,7 +366,7 @@ async fn grpc_internal() {
 async fn grpc_timeout() {
     let _trace = trace::test::with_default_filter("linkerd=debug");
 
-    const TIMEOUT: Duration = Duration::from_millis(100);
+    const TIMEOUT: time::Duration = time::Duration::from_millis(100);
     let (svc, mut handle) = mock_grpc(GrpcParams {
         timeouts: Timeouts {
             request: Some(TIMEOUT * 5),
@@ -378,14 +386,13 @@ async fn grpc_timeout() {
         async move {
             handle.allow(2);
             info!("Delaying the first request");
-            serve_delayed(
-                TIMEOUT * 2,
-                &mut handle,
-                Ok(mk_grpc_rsp(tonic::Code::NotFound)),
-            )
+            serve(&mut handle, async move {
+                time::sleep(TIMEOUT * 2).await;
+                mk_grpc_rsp(tonic::Code::NotFound).await
+            })
             .await;
             info!("Serving the second request");
-            serve_req(&mut handle, mk_grpc_rsp(tonic::Code::Ok)).await;
+            serve(&mut handle, mk_grpc_rsp(tonic::Code::Ok)).await;
             handle
         }
         .in_current_span(),
@@ -417,89 +424,4 @@ async fn grpc_timeout() {
             .unwrap(),
         "0"
     );
-}
-
-// === Utils ===
-
-type Handle = tower_test::mock::Handle<Request, Response>;
-
-fn mock_http(params: HttpParams) -> (svc::BoxCloneHttp, Handle) {
-    let dest: NameAddr = "example.com:1234".parse::<NameAddr>().unwrap();
-    let backend = default_backend(&dest);
-    let route = mk_route(backend.clone(), params);
-    mock(
-        dest.clone(),
-        policy::Params::Http(policy::HttpParams {
-            addr: dest.into(),
-            meta: ParentRef(client_policy::Meta::new_default("parent")),
-            backends: Arc::new([backend]),
-            routes: Arc::new([route]),
-            failure_accrual: client_policy::FailureAccrual::None,
-        }),
-    )
-}
-
-fn mock_grpc(params: GrpcParams) -> (svc::BoxCloneHttp, Handle) {
-    let dest: NameAddr = "example.com:1234".parse::<NameAddr>().unwrap();
-    let backend = default_backend(&dest);
-    let route = mk_route(backend.clone(), params);
-    mock(
-        dest.clone(),
-        policy::Params::Grpc(policy::GrpcParams {
-            addr: dest.into(),
-            meta: ParentRef(client_policy::Meta::new_default("parent")),
-            backends: Arc::new([backend]),
-            routes: Arc::new([route]),
-            failure_accrual: client_policy::FailureAccrual::None,
-        }),
-    )
-}
-
-fn mock(dest: NameAddr, params: policy::Params) -> (svc::BoxCloneHttp, Handle) {
-    let addr = SocketAddr::new([192, 0, 2, 41].into(), 1234);
-
-    let (inner, handle) = tower_test::mock::pair();
-    let connect = HttpConnect::default().service(addr, inner);
-    let resolve = support::resolver().endpoint_exists(dest.clone(), addr, Default::default());
-    let (rt, _shutdown) = runtime();
-    let stack = Outbound::new(default_config(), rt, &mut Default::default())
-        .with_stack(svc::ArcNewService::new(connect))
-        .push_http_cached(resolve)
-        .into_inner();
-
-    let (tx, routes) = watch::channel(Routes::Policy(params));
-    tokio::spawn(async move {
-        tx.closed().await;
-    });
-
-    let svc = stack.new_service(Target {
-        num: 1,
-        version: http::Version::H2,
-        routes,
-    });
-
-    (svc, handle)
-}
-
-fn mk_route<M: Default, F, P>(
-    backend: client_policy::Backend,
-    params: P,
-) -> client_policy::route::Route<M, client_policy::RoutePolicy<F, P>> {
-    use client_policy::*;
-
-    route::Route {
-        hosts: vec![],
-        rules: vec![route::Rule {
-            matches: vec![M::default()],
-            policy: RoutePolicy {
-                meta: Meta::new_default("route"),
-                filters: [].into(),
-                distribution: RouteDistribution::FirstAvailable(Arc::new([RouteBackend {
-                    filters: [].into(),
-                    backend,
-                }])),
-                params,
-            },
-        }],
-    }
 }
