+++ linkerd/app/outbound/src/http/logical/tests/timeouts.rs
@@ -1,42 +1,40 @@
-use super::{
-    super::{policy, LogicalError, Outbound, ParentRef, Routes},
-    default_backend, http_get, mk_rsp, send_req, serve_delayed, serve_req, HttpConnect, Request,
-    Response, Target,
-};
-use crate::test_util::*;
+use super::{super::LogicalError, *};
 use linkerd_app_core::{
     errors,
     proxy::http::{
         self,
         stream_timeouts::{BodyTimeoutError, ResponseTimeoutError},
-        BoxBody, HttpBody, StatusCode,
+        BoxBody, HttpBody,
     },
-    svc::{self, NewService},
-    trace, NameAddr,
+    trace,
 };
 use linkerd_proxy_client_policy::{self as client_policy, http::Timeouts};
-use std::{net::SocketAddr, sync::Arc, time::Duration};
-use tokio::{sync::watch, time};
-use tracing::info;
+use tokio::time;
+use tracing::{info, Instrument};
 
 #[tokio::test(flavor = "current_thread", start_paused = true)]
 async fn request_timeout_response_headers() {
     let _trace = trace::test::trace_init();
 
-    const TIMEOUT: Duration = Duration::from_secs(2);
-    let (svc, mut handle) = mock(Timeouts {
-        request: Some(TIMEOUT),
+    const TIMEOUT: time::Duration = time::Duration::from_secs(2);
+    let (svc, mut handle) = mock_http(client_policy::http::RouteParams {
+        timeouts: Timeouts {
+            request: Some(TIMEOUT),
+            ..Default::default()
+        },
         ..Default::default()
     });
 
     info!("Sending a request that does not respond within the timeout");
     handle.allow(1);
     let call = send_req(svc.clone(), http_get());
-    serve_delayed(
-        TIMEOUT * 2,
-        &mut handle,
-        Ok(mk_rsp(StatusCode::NO_CONTENT, "")),
-    )
+    serve(&mut handle, async move {
+        time::sleep(TIMEOUT * 2).await;
+        Ok(http::Response::builder()
+            .status(204)
+            .body(http::BoxBody::default())
+            .unwrap())
+    })
     .await;
 
     info!("Verifying that the response fails with the expected error");
@@ -61,9 +59,12 @@ async fn request_timeout_response_headers() {
 async fn request_timeout_request_body() {
     let _trace = trace::test::trace_init();
 
-    const TIMEOUT: Duration = Duration::from_secs(2);
-    let (svc, mut handle) = mock(Timeouts {
-        request: Some(TIMEOUT),
+    const TIMEOUT: time::Duration = time::Duration::from_secs(2);
+    let (svc, mut handle) = mock_http(client_policy::http::RouteParams {
+        timeouts: Timeouts {
+            request: Some(TIMEOUT),
+            ..Default::default()
+        },
         ..Default::default()
     });
 
@@ -101,23 +102,28 @@ async fn request_timeout_request_body() {
 async fn request_timeout_response_body() {
     let _trace = trace::test::trace_init();
 
-    const TIMEOUT: Duration = Duration::from_secs(2);
-    let (svc, mut handle) = mock(Timeouts {
-        request: Some(TIMEOUT),
+    const TIMEOUT: time::Duration = time::Duration::from_secs(2);
+    let (svc, mut handle) = mock_http(client_policy::http::RouteParams {
+        timeouts: Timeouts {
+            request: Some(TIMEOUT),
+            ..Default::default()
+        },
         ..Default::default()
     });
 
     info!("Sending a request that responds immediately but does not complete");
     handle.allow(1);
     let call = send_req(svc.clone(), http_get());
-    serve_req(
+    serve(
         &mut handle,
-        http::Response::builder()
-            .status(200)
-            .body(http::BoxBody::new(MockBody::new(async move {
-                futures::future::pending().await
-            })))
-            .unwrap(),
+        future::ok(
+            http::Response::builder()
+                .status(200)
+                .body(http::BoxBody::new(MockBody::new(async move {
+                    futures::future::pending().await
+                })))
+                .unwrap(),
+        ),
     )
     .await;
 
@@ -142,23 +148,25 @@ async fn request_timeout_response_body() {
 async fn response_timeout_response_headers() {
     let _trace = trace::test::with_default_filter("linkerd=trace");
 
-    const TIMEOUT: Duration = Duration::from_secs(2);
-    let (svc, mut handle) = mock(Timeouts {
-        response: Some(TIMEOUT),
+    const TIMEOUT: time::Duration = time::Duration::from_secs(2);
+    let (svc, mut handle) = mock_http(client_policy::http::RouteParams {
+        timeouts: Timeouts {
+            response: Some(TIMEOUT),
+            ..Default::default()
+        },
         ..Default::default()
     });
 
     info!("Sending a request that does not respond within the timeout");
     handle.allow(1);
     let call = send_req(svc.clone(), http_get());
-    serve_delayed(
-        TIMEOUT * 2,
-        &mut handle,
+    serve(&mut handle, async move {
+        time::sleep(TIMEOUT * 2).await;
         Ok(http::Response::builder()
             .status(204)
             .body(http::BoxBody::default())
-            .unwrap()),
-    )
+            .unwrap())
+    })
     .await;
 
     info!("Verifying that the response fails with the expected error");
@@ -179,28 +187,36 @@ async fn response_timeout_response_headers() {
 async fn response_timeout_response_body() {
     let _trace = trace::test::trace_init();
 
-    const TIMEOUT: Duration = Duration::from_secs(2);
-    let (svc, mut handle) = mock(Timeouts {
-        response: Some(TIMEOUT),
+    const TIMEOUT: time::Duration = time::Duration::from_secs(2);
+    let (svc, mut handle) = mock_http(client_policy::http::RouteParams {
+        timeouts: Timeouts {
+            response: Some(TIMEOUT),
+            ..Default::default()
+        },
         ..Default::default()
     });
 
+    tokio::spawn(
+        async move {
+            handle.allow(1);
+            serve(&mut handle, async move {
+                info!("Serving a response that never completes");
+                Ok(http::Response::builder()
+                    .status(200)
+                    .body(http::BoxBody::new(MockBody::new(async move {
+                        futures::future::pending().await
+                    })))
+                    .unwrap())
+            })
+            .await;
+        }
+        .in_current_span(),
+    );
+
     info!("Sending a request that responds immediately but does not complete");
-    handle.allow(1);
-    let call = send_req(svc.clone(), http_get());
-    serve_req(
-        &mut handle,
-        http::Response::builder()
-            .status(200)
-            .body(http::BoxBody::new(MockBody::new(async move {
-                futures::future::pending().await
-            })))
-            .unwrap(),
-    )
-    .await;
+    let mut rsp = send_req(svc.clone(), http_get()).await.unwrap().into_body();
 
     info!("Verifying that the request body times out with the expected stream error");
-    let mut rsp = call.await.unwrap().into_body();
     let error = time::timeout(TIMEOUT * 2, rsp.data())
         .await
         .expect("should timeout internally")
@@ -220,25 +236,27 @@ async fn response_timeout_response_body() {
 async fn response_timeout_ignores_request_body() {
     let _trace = trace::test::with_default_filter("linkerd=trace");
 
-    const TIMEOUT: Duration = Duration::from_secs(2);
-    let (svc, mut handle) = mock(Timeouts {
-        response: Some(TIMEOUT),
+    const TIMEOUT: time::Duration = time::Duration::from_secs(2);
+    let (svc, mut handle) = mock_http(client_policy::http::RouteParams {
+        timeouts: Timeouts {
+            response: Some(TIMEOUT),
+            ..Default::default()
+        },
         ..Default::default()
     });
 
     info!("Sending a request that exceeds the response timeout");
     handle.allow(1);
     let call = send_req(svc.clone(), http_get());
-    serve_req(
-        &mut handle,
-        http::Response::builder()
+    serve(&mut handle, async move {
+        info!("Serving a response that never completes");
+        Ok(http::Response::builder()
             .status(200)
             .body(http::BoxBody::new(MockBody::new(async move {
-                time::sleep(TIMEOUT * 2).await;
-                Ok(())
+                futures::future::pending().await
             })))
-            .unwrap(),
-    )
+            .unwrap())
+    })
     .await;
 
     info!("Verifying that the response succeeds despite slow request time");
@@ -252,24 +270,27 @@ async fn response_timeout_ignores_request_body() {
 async fn idle_timeout_response_body() {
     let _trace = trace::test::trace_init();
 
-    const TIMEOUT: Duration = Duration::from_secs(2);
-    let (svc, mut handle) = mock(Timeouts {
-        idle: Some(TIMEOUT),
+    const TIMEOUT: time::Duration = time::Duration::from_secs(2);
+    let (svc, mut handle) = mock_http(client_policy::http::RouteParams {
+        timeouts: Timeouts {
+            idle: Some(TIMEOUT),
+            ..Default::default()
+        },
         ..Default::default()
     });
 
     info!("Sending a request that is served immediately with a body that does not update");
     handle.allow(1);
     let call = send_req(svc.clone(), http_get());
-    serve_req(
-        &mut handle,
-        http::Response::builder()
+    serve(&mut handle, async move {
+        info!("Serving a response that never completes");
+        Ok(http::Response::builder()
             .status(200)
             .body(http::BoxBody::new(MockBody::new(async move {
                 futures::future::pending().await
             })))
-            .unwrap(),
-    )
+            .unwrap())
+    })
     .await;
 
     info!("Verifying that the request body times out with the expected stream error");
@@ -287,70 +308,4 @@ async fn idle_timeout_response_body() {
         ),
         "expected idle timeout, got {error:?}"
     );
-}
-
-// === Utils ===
-
-type Handle = tower_test::mock::Handle<Request, Response>;
-
-fn mock(timeouts: Timeouts) -> (svc::BoxCloneHttp, Handle) {
-    let addr = SocketAddr::new([192, 0, 2, 41].into(), 1234);
-    let dest: NameAddr = "example.com:1234".parse::<NameAddr>().unwrap();
-
-    let (inner, handle) = tower_test::mock::pair();
-    let connect = HttpConnect::default().service(addr, inner);
-    let resolve = support::resolver().endpoint_exists(dest.clone(), addr, Default::default());
-    let (rt, _shutdown) = runtime();
-    let stack = Outbound::new(default_config(), rt, &mut Default::default())
-        .with_stack(svc::ArcNewService::new(connect))
-        .push_http_cached(resolve)
-        .into_inner();
-
-    let (tx, routes) = {
-        let backend = default_backend(&dest);
-        let route = mk_route(backend.clone(), timeouts);
-        watch::channel(Routes::Policy(policy::Params::Http(policy::HttpParams {
-            addr: dest.into(),
-            meta: ParentRef(client_policy::Meta::new_default("parent")),
-            backends: Arc::new([backend]),
-            routes: Arc::new([route]),
-            failure_accrual: client_policy::FailureAccrual::None,
-        })))
-    };
-    tokio::spawn(async move {
-        tx.closed().await;
-    });
-
-    let svc = stack.new_service(Target {
-        num: 1,
-        version: http::Version::H2,
-        routes,
-    });
-
-    (svc, handle)
-}
-
-fn mk_route(backend: client_policy::Backend, timeouts: Timeouts) -> client_policy::http::Route {
-    use client_policy::{
-        http::{self, Policy, Route, Rule},
-        Meta, RouteBackend, RouteDistribution,
-    };
-    Route {
-        hosts: vec![],
-        rules: vec![Rule {
-            matches: vec![http::r#match::MatchRequest::default()],
-            policy: Policy {
-                meta: Meta::new_default("timeout-route"),
-                filters: [].into(),
-                params: http::RouteParams {
-                    timeouts,
-                    ..Default::default()
-                },
-                distribution: RouteDistribution::FirstAvailable(Arc::new([RouteBackend {
-                    filters: [].into(),
-                    backend,
-                }])),
-            },
-        }],
-    }
 }
